SuiteClient::create() accepts empty sourceId
